# Support AVX-512 for on demand parsing