… cmake/set_arch_flags.cmake
… include/sonic/internal/arch/avx512/simd.h