Preliminary support for AVX-512 for on demand parsing.

This was tested on a Ryzen5 AMD EPYC 9755 processor.

Although the current performance is slightly slower than regular AVX2 due to `_mm512_cmpeq_epi8_mask` low throughput and high latency in the tested system., future processors may benefit from having AVX-512 support already included.

To enable it, compile with `-DENABLE_AVX_512`

Master
```
twitter/SonicOnDemand_Normal            10066 ns        10066 ns        68964 bytes_per_second=24.4606Gi/s Normal
citm_catalog/SonicOnDemand_Fronter       6320 ns         6320 ns       111994 bytes_per_second=254.51Gi/s Fronter
twitter/SonicOnDemand_NotFound           9986 ns         9986 ns        69888 bytes_per_second=24.6569Gi/s NotFound
```

AVX-512

```
twitter/SonicOnDemand_Normal            10389 ns        10389 ns        67310 bytes_per_second=23.6997Gi/s Normal
citm_catalog/SonicOnDemand_Fronter       7943 ns         7943 ns        89128 bytes_per_second=202.526Gi/s Fronter
twitter/SonicOnDemand_NotFound          10259 ns        10259 ns        67899 bytes_per_second=24.0019Gi/s NotFound
```

Thanks to @supermartian 

This PR is contributed by nvidia